Extracting encryption from cdap-security into cdap-common
