+++ cdap-common/src/main/java/io/cdap/cdap/common/encryption/NoOpAeadCipher.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package io.cdap.cdap.security.encryption;
+package io.cdap.cdap.common.encryption;
 
 /**
  * Performs no encryption.
