+++ cdap-common/src/main/java/io/cdap/cdap/common/encryption/guice/LazyDelegateAeadCipher.java
@@ -14,9 +14,9 @@
  * the License.
  */
 
-package io.cdap.cdap.security.encryption.guice;
+package io.cdap.cdap.common.encryption.guice;
 
-import io.cdap.cdap.security.encryption.AeadCipher;
+import io.cdap.cdap.common.encryption.AeadCipher;
 import io.cdap.cdap.security.spi.encryption.AeadCipherContext;
 import io.cdap.cdap.security.spi.encryption.AeadCipherCryptor;
 import io.cdap.cdap.security.spi.encryption.CipherException;
