+++ cdap-gateway/src/main/java/io/cdap/cdap/gateway/router/handlers/AuthenticationHandler.java
@@ -24,6 +24,7 @@ import io.cdap.cdap.common.conf.CConfiguration;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.conf.Constants.Security.Encryption;
 import io.cdap.cdap.common.conf.SConfiguration;
+import io.cdap.cdap.common.encryption.AeadCipher;
 import io.cdap.cdap.common.logging.AuditLogEntry;
 import io.cdap.cdap.common.utils.Networks;
 import io.cdap.cdap.proto.security.Credential;
@@ -32,7 +33,6 @@ import io.cdap.cdap.security.auth.UserIdentityExtractionResponse;
 import io.cdap.cdap.security.auth.UserIdentityExtractionState;
 import io.cdap.cdap.security.auth.UserIdentityExtractor;
 import io.cdap.cdap.security.auth.UserIdentityPair;
-import io.cdap.cdap.security.encryption.AeadCipher;
 import io.cdap.cdap.security.server.GrantAccessToken;
 import io.cdap.cdap.security.spi.encryption.CipherException;
 import io.netty.buffer.ByteBuf;
