+++ cdap-gateway/src/test/java/io/cdap/cdap/gateway/router/AuditLogTest.java
@@ -28,10 +28,10 @@ import com.google.common.io.ByteStreams;
 import io.cdap.cdap.common.conf.CConfiguration;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.conf.SConfiguration;
+import io.cdap.cdap.common.encryption.NoOpAeadCipher;
 import io.cdap.cdap.common.security.AuditDetail;
 import io.cdap.cdap.common.security.AuditPolicy;
 import io.cdap.cdap.security.auth.TokenValidator;
-import io.cdap.cdap.security.encryption.NoOpAeadCipher;
 import io.cdap.http.AbstractHttpHandler;
 import io.cdap.http.HttpResponder;
 import io.cdap.http.NettyHttpService;
