+++ cdap-gateway/src/test/java/io/cdap/cdap/gateway/router/NettyRouterTestBase.java
@@ -36,9 +36,9 @@ import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.conf.SConfiguration;
 import io.cdap.cdap.common.discovery.RandomEndpointStrategy;
 import io.cdap.cdap.common.discovery.ResolvingDiscoverable;
+import io.cdap.cdap.common.encryption.NoOpAeadCipher;
 import io.cdap.cdap.common.http.AbstractBodyConsumer;
 import io.cdap.cdap.security.auth.TokenValidator;
-import io.cdap.cdap.security.encryption.NoOpAeadCipher;
 import io.cdap.http.AbstractHttpHandler;
 import io.cdap.http.BodyConsumer;
 import io.cdap.http.ChannelPipelineModifier;
