+++ cdap-security/src/test/java/io/cdap/cdap/security/authorization/DefaultAccessEnforcerTest.java
@@ -31,6 +31,7 @@ import io.cdap.cdap.common.conf.CConfiguration;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.conf.Constants.Security.Encryption;
 import io.cdap.cdap.common.conf.SConfiguration;
+import io.cdap.cdap.common.encryption.FakeAeadCipher;
 import io.cdap.cdap.common.metrics.ProgramTypeMetricTag;
 import io.cdap.cdap.common.test.AppJarHelper;
 import io.cdap.cdap.proto.ProgramType;
@@ -47,7 +48,6 @@ import io.cdap.cdap.proto.security.Principal;
 import io.cdap.cdap.proto.security.StandardPermission;
 import io.cdap.cdap.security.auth.CipherException;
 import io.cdap.cdap.security.auth.context.AuthenticationTestContext;
-import io.cdap.cdap.security.encryption.FakeAeadCipher;
 import io.cdap.cdap.security.spi.authentication.AuthenticationContext;
 import io.cdap.cdap.security.spi.authorization.AccessControllerSpi;
 import io.cdap.cdap.security.spi.authorization.AccessEnforcer;
